Importing H2 DWave Annealer
Imported the Week 1 H2 DWave Annealer code along with the energies table. Will need to adapt for easy insert into or alongside Task 3 for comparisons. There is a error in the energies table that is mitigated in the code. This has already been reviewed and confirmed by ejazz at CDL.